Quoted keys/values are unquoted
Single-quotes can be used to escape certain characters (such as `,` and `:`)
but their quotation is still kept. This causes `go-pg` to call `Unquote`
compulsively and also fails the mapping in `msgpack` when single-quotes
are used to escape a custom-fieldname.

Resolve this by removing the single-quotes from quoted values unless
they've been escaped with a backslash.

Unquote is removed because it has no purpose anymore.